Fix Memory Corruption Unhandled Memory fault at #x0
It seems that `rl:redisplay` initializes  some C pointer which is not
properly setup when is overridden by `enable-syntax`, that way a
SEGVFAULT signal raises and the signal is handled before. Using
`strace ros repl` it's possible to check that.

Calling `(rl:redisplay)` on the beginning of
`redisplay-with-highlight` function, which is used
when (enable-syntax) is called, fix that issue.

That is some mistery to me yet, I look directly on cl-readline and GNU
readline C code but it's not clear what is going here.

Proof of working:

![fix-of-the-fix](https://imgur.com/ELfCTdc.png)

Closes #44 and #50